refactored hex block and the components that used it
the hex block component was initially coupled to be used as a hoc on a per component
basis. This coupling made it necessary to pass props down to the hoc that were irrelevant to the
child component that was being wrapped be the hex block. The component has been refactored into being
owned by a Hexgrid component which is responsible for wrapping any items that go into it with hex blocks
this allows for a better interface when passing props to 'configure' the hexblocks